+++ package.json
@@ -34,7 +34,7 @@
     "@typescript-eslint/parser": "^5.53.0",
     "autoprefixer": "^10.4.7",
     "eslint": "^8.34.0",
-    "eslint-config-next": "13.1.6",
+    "eslint-config-next": "13.2.1",
     "postcss": "^8.4.14",
     "prettier": "^2.8.1",
     "prettier-plugin-tailwindcss": "^0.2.1",
