+++ src/app/layout.tsx
@@ -1,0 +1,27 @@
+import "./globals.css";
+
+export const metadata = {
+  title: "Create Next App",
+  description: "Generated by create next app",
+};
+
+import { ClerkProvider } from "@clerk/nextjs/app-beta";
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  console.log("made it to layout");
+  return (
+    <html lang="en">
+      <ClerkProvider>
+        <body>
+          <main className="flex min-h-screen flex-col items-center bg-black text-white">
+            {children}
+          </main>
+        </body>
+      </ClerkProvider>
+    </html>
+  );
+}
