--- src/pages/_app.tsx
@@ -1,48 +1,0 @@
-import {
-  ClerkProvider,
-  SignedIn,
-  SignedOut,
-  RedirectToSignIn,
-} from "@clerk/nextjs";
-import { AppProps } from "next/app";
-import { useRouter } from "next/router";
-import "~/styles/globals.css";
-
-//  List pages you want to be publicly accessible, or leave empty if
-//  every page requires authentication. Use this naming strategy:
-//   "/"              for pages/index.js
-//   "/foo"           for pages/foo/index.js
-//   "/foo/bar"       for pages/foo/bar.js
-//   "/foo/[...bar]"  for pages/foo/[...bar].js
-const publicPages: Array<string> = ["/", "/about"];
-
-function MainApp({ Component, pageProps }: AppProps) {
-  // Get the pathname
-  const { pathname } = useRouter();
-
-  // Check if the current route matches a public page
-  const isPublicPage = publicPages.includes(pathname);
-
-  // If the current route is listed as public, render it directly
-  // Otherwise, use Clerk to require authentication
-  return (
-    <ClerkProvider>
-      {isPublicPage ? (
-        <Component {...pageProps} />
-      ) : (
-        <>
-          <SignedIn>
-            <Component {...pageProps} />
-          </SignedIn>
-          <SignedOut>
-            <RedirectToSignIn />
-          </SignedOut>
-        </>
-      )}
-    </ClerkProvider>
-  );
-}
-
-// Wrap with tRPC
-import { api } from "~/utils/api";
-export default api.withTRPC(MainApp);
