--- src/pages/index.tsx
@@ -1,104 +1,0 @@
-import { useUser } from "@clerk/nextjs";
-import { type NextPage } from "next";
-import Head from "next/head";
-import Link from "next/link";
-import { useState } from "react";
-import dayjs from "dayjs";
-import relativeTime from "dayjs/plugin/relativeTime";
-
-dayjs.extend(relativeTime);
-
-import { api, RouterOutputs } from "~/utils/api";
-
-type TweetData = RouterOutputs["example"]["getAll"][number];
-
-const TweetView = (props: { tweet: TweetData }) => {
-  return (
-    <div className="border-t border-zinc-700 p-4 shadow-lg">
-      <Link href={`/post/${props.tweet.id}`}>
-        <div className="flex items-center">
-          <img
-            src={props.tweet.user.profileImageUrl}
-            alt="Profile"
-            className="h-14 w-14 rounded-full"
-          />
-          <div className="ml-3 flex flex-col text-2xl">
-            <div className="text-base font-bold text-slate-300">
-              <span>{`@${props.tweet.user.username}`}</span>
-              <span className="font-thin">{` · ${dayjs(
-                props.tweet.createdAt
-              ).fromNow()}`}</span>
-            </div>
-            <div className="text-slate-300">{props.tweet.content}</div>
-          </div>
-        </div>
-      </Link>
-    </div>
-  );
-};
-
-const CreatePostWizard = () => {
-  const [content, setContent] = useState("");
-
-  const ctx = api.useContext();
-  const { mutate, isLoading } = api.example.createPost.useMutation({
-    onSuccess: () => {
-      setContent("");
-      ctx.invalidate();
-    },
-  });
-
-  const { user } = useUser();
-
-  return (
-    <div className="relative flex w-full">
-      <img
-        src={user?.profileImageUrl}
-        alt="Profile"
-        className="m-4 h-14 w-14 rounded-full"
-      />
-      <input
-        value={content}
-        onChange={(e) => setContent(e.target.value)}
-        disabled={isLoading}
-        className="my-4 grow bg-transparent py-4 pr-20 text-xl outline-none"
-        placeholder="Type some emojis"
-        autoFocus
-      />
-      <div className="absolute right-2 flex h-full flex-col justify-center">
-        {!!content && (
-          <button onClick={() => mutate({ message: content })}>POST!</button>
-        )}
-      </div>
-    </div>
-  );
-};
-
-const Feed = () => {
-  const { data } = api.example.getAll.useQuery();
-  return (
-    <div className="flex h-full w-full grow flex-col border-l border-r border-zinc-700 md:w-[600px]">
-      <CreatePostWizard />
-      {data?.map((post) => (
-        <TweetView key={post.id} tweet={post} />
-      ))}
-    </div>
-  );
-};
-
-const Home: NextPage = () => {
-  return (
-    <>
-      <Head>
-        <title>😶 Emojer</title>
-        <meta name="description" content="Generated by create-t3-app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <main className="flex min-h-screen flex-col items-center bg-black text-white">
-        <Feed />
-      </main>
-    </>
-  );
-};
-
-export default Home;
