--- src/pages/post/[id].tsx
@@ -1,67 +1,0 @@
-import { type NextPage } from "next";
-import Head from "next/head";
-import dayjs from "dayjs";
-import relativeTime from "dayjs/plugin/relativeTime";
-
-dayjs.extend(relativeTime);
-
-import { api, RouterOutputs } from "~/utils/api";
-import { useRouter } from "next/router";
-
-type TweetData = RouterOutputs["example"]["getAll"][number];
-
-const TweetView = (props: { tweet: TweetData }) => {
-  return (
-    <div className="border-t border-zinc-700 p-4 shadow-lg">
-      <div className="flex items-center">
-        <img
-          src={props.tweet.user.profileImageUrl}
-          alt="Profile"
-          className="h-14 w-14 rounded-full"
-        />
-        <div className="ml-3 flex flex-col text-2xl">
-          <div className="text-base font-bold text-slate-300">
-            <span>{`@${props.tweet.user.username}`}</span>
-            <span className="font-thin">{` · ${dayjs(
-              props.tweet.createdAt
-            ).fromNow()}`}</span>
-          </div>
-          <div className="text-slate-300">{props.tweet.content}</div>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-const Feed = (props: { id: string }) => {
-  const { data } = api.example.getPostById.useQuery({ id: props.id });
-  return (
-    <div className="flex h-full w-full grow flex-col border-l border-r border-zinc-700 md:w-[600px]">
-      {data && <TweetView tweet={data} />}
-    </div>
-  );
-};
-
-const PostView: NextPage = () => {
-  const { query } = useRouter();
-  console.log("router", query);
-
-  if (!query.id || typeof query.id !== "string") {
-    return <div>Not found</div>;
-  }
-
-  return (
-    <>
-      <Head>
-        <title>😶 Post</title>
-        <meta name="description" content="Generated by create-t3-app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <main className="flex min-h-screen flex-col items-center bg-black text-white">
-        <Feed id={query.id} />
-      </main>
-    </>
-  );
-};
-
-export default PostView;
